updated profiles.js for profiles.d
